forgive me linus for I have sinned, it has been too long since my last commit.
removed description from recipe, add unit dropdown in order to normalize unit inputs, add ingredientListItemRepository, change listitem quantlty from int to float

+++ app/src/main/java/dk/arongk/and1_recipeapp/models/ingredientListItem/IngredientListItemCreateModel.kt
@@ -3,7 +3,7 @@ package dk.arongk.and1_recipeapp.models.ingredientListItem
 import java.util.*
 
 data class IngredientListItemCreateModel(
-    var quantity: Int,
+    var quantity: Float,
     var ingredientName: String,
     var unit: String,
     var operation: String
